Call sinon with proper arguments (fixes failing txn tests)
Jasmine seems to allow for named anonymous spy functions. Sinon does not. sinon.spy accepts obj/fn, method. If there are 2 arguments, it assumes the first is an object and the second is a method name. For one argument, it assumes a function to wrap and return. I've updated the spyFactory to just ignore the names if sinon is in use.