allow specifying additional user attributes
Currently, `createUser` only allows specifying uid, email and password.
As the real `firebase.User` interface has many more properties (ie.
`displayName`, `disabled`, ...), some tests are not easy to configure in
firebase-mock.

To help here, copy all properties of credentials over to the newly created user object.
A whitelist of allowed properties could be added as well as validation
rules on types and values, but this would increase maintenance as it
would need to be kept up-to-date.

Docs:
https://firebase.google.com/docs/reference/js/firebase.User